Added pomodoro context for sharing data between our timer and task components
After completing a pomodoro cycle, we need to update items in our tasks component such as pomodoro counts for each tasks -- we need to mark theses tasks as completed and change the active task. As both components modify or require this information, a context is needed to share it between them.